Services, Jury, Admin

+++ src/main/java/devlight/edu/conference/service/ApplicationServiceImpl.java
@@ -1,34 +1,20 @@
 package devlight.edu.conference.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import devlight.edu.conference.model.Application;
 import devlight.edu.conference.repository.ApplicationRepository;
-import devlight.edu.conference.repository.CuratorRepository;
-import devlight.edu.conference.repository.DirectionRepository;
+import javassist.NotFoundException;
 
 @Service
 public class ApplicationServiceImpl implements ApplicationService {
 
 	@Autowired
 	ApplicationRepository applicationRepository;
-	@Autowired
-	CuratorRepository curatorRepository;
-	@Autowired
-	DirectionRepository directionRepository;
-
-	@Override
-	public Application getApplicationById(int id) {
-		return applicationRepository.getOne(id);
-	}
-
-	@Override
-	public List<Application> getAllApplications() {
-		return applicationRepository.findAll();
-	}
 
 	@Override
 	public Application addApplication(Application application) {
@@ -36,14 +22,31 @@ public class ApplicationServiceImpl implements ApplicationService {
 	}
 
 	@Override
-	public void deleteApplication(int id) {
-		applicationRepository.deleteById(id);
-
-	}
-
-	@Override
-	public void updateApplication(Application application) {
+	public void editApplication(Application application) {
 		applicationRepository.save(application);
 	}
 
+	@Override
+	public Application getApplicationById(int id) throws NotFoundException {
+		Optional<Application> application = applicationRepository.findById(id);
+		if (application.isPresent())
+			return application.get();
+		else
+			throw new NotFoundException("Application is not found");
+	}
+
+	@Override
+	public List<Application> getAllApplications() throws NotFoundException {
+		List<Application> applicationList = applicationRepository.findAll();
+		if (applicationList != null)
+			return applicationList;
+		else
+			throw new NotFoundException("There are no any applications in DataBase");
+	}
+
+	@Override
+	public void deleteApplication(int id) throws NotFoundException {
+		if (getApplicationById(id) != null)
+			applicationRepository.deleteById(id);
+	}
 }
